powerpc/selftests: add test for papr-vpd
Add test coverage for the common case of retrieving all VPD by passing
an empty location code string, as well as verifying behavior on bad
inputs.

Not covered currently: retrieving a subset of VPD by passing a
specific location code. This is possible, but believe we'd have to
discover a suitable location code at runtime and it would differ
between systems.

The tests are skipped on systems where the papr-vpd driver does not
initialize, making this useful only on PowerVM LPARs at this point.

Signed-off-by: Nathan Lynch <nathanl@linux.ibm.com>