powerpc/smp: add freeze/thaw functions to smp_ops
These are platform-specific hooks for pausing and resuming CPUs during
suspend. They are meant to be lighter weight and faster than offlining
and onlining CPUs, which is what the suspend core does now.

The only expected user is pseries, which will use the H_JOIN and
H_PROD hcalls to provide the desired semantics.

Signed-off-by: Nathan Lynch <nathanl@linux.ibm.com>